+++ packages/react/src/format-countdown.ts
@@ -1,0 +1,11 @@
+export const formatCountdown = (secondsLeft: number): string => {
+  const total = Math.max(0, Math.floor(secondsLeft));
+  const days = Math.floor(total / 86400);
+  const hours = Math.floor((total % 86400) / 3600);
+  const minutes = Math.floor((total % 3600) / 60);
+  const seconds = total % 60;
+  const hh = hours.toString().padStart(2, "0");
+  const mm = minutes.toString().padStart(2, "0");
+  const ss = seconds.toString().padStart(2, "0");
+  return days > 0 ? `${days}d ${hh}:${mm}:${ss}` : `${hh}:${mm}:${ss}`;
+};
